Fallos de seguridad arreglados

+++ ProyectoFinal/app/Http/Controllers/RolController.php
@@ -6,6 +6,7 @@ use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Foundation\Bus\DispatchesJobs;
 use Illuminate\Foundation\Validation\ValidatesRequests;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 use App\Models\Rol;
 class RolController extends Controller
@@ -14,37 +15,53 @@ class RolController extends Controller
 
     function list()
     {
-        $rols = Rol::all();
-        return view('rol.list', ['rols'=>$rols]);
+        if (Auth::user()->rol_id == 5076) {
+            $rols = Rol::all();
+            return view('rol.list', ['rols'=>$rols]);
+        } else {
+          return redirect('');
+        }
     }
 
     function new (Request $request)
     {
-        if ($request->isMethod('post')){
-            $rol = new Rol;
-            $rol->name = $request->name;
-            $rol->save();
-            return redirect()->route('rol_list');
+        if (Auth::user()->rol_id == 5076) {
+            if ($request->isMethod('post')){
+                $rol = new Rol;
+                $rol->name = $request->name;
+                $rol->save();
+                return redirect()->route('rol_list');
+            }
+            $rols = Rol::all();
+            return view('rol.new', ['rols' => $rols]);
+        } else {
+        return redirect('');
         }
-        $rols = Rol::all();
-        return view('rol.new', ['rols' => $rols]);
     }
 
     function edit(Request $request, $id)
     {
-        $rol = Rol::find($id);
-        if($request->isMethod('post')){
-            $rol->name = $request->name;
-            $rol->save();
-            return redirect()->route('user_list');
+        if (Auth::user()->rol_id == 5076) {
+            $rol = Rol::find($id);
+            if($request->isMethod('post')){
+                $rol->name = $request->name;
+                $rol->save();
+                return redirect()->route('user_list');
+            }
+            $rols = Rol::all();
+            return view('rol.edit', ['rols'=>$rols, 'rol'=>$rol]);
+        } else {
+        return redirect('');
         }
-        $rols = Rol::all();
-        return view('rol.edit', ['rols'=>$rols, 'rol'=>$rol]);
     }
     function delete($id)
     {
-        $rol = Rol::find($id);
-        $rol->delete();
-        return redirect()->route('rol_list');
+        if (Auth::user()->rol_id == 5076) {
+            $rol = Rol::find($id);
+            $rol->delete();
+            return redirect()->route('rol_list');
+        } else {
+        return redirect('');
+        }
     }
 }
